src/main/java/sk/umb/example/mockito/BookController.java
package sk.umb.example.mockito;

import lombok.RequiredArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequiredArgsConstructor
public class BookController {
    @Autowired
    private BookService bookService;

    @PostMapping("/api/books")
    public ResponseEntity<Long> saveBook(@RequestBody BookDto bookDto) {
        Long id = bookService.saveBook(bookDto);

        return new ResponseEntity<>(id, HttpStatus.CREATED);
    }

    @GetMapping("/api/books")
    public List<BookDto> getAllBooks() {
        return bookService.getAllBooks();
    }
}
